refactor: Relocate SaturationDetector
This commit moves the existing `SaturationDetector` implementation to
its new home under the EPP plugin framework structure. This is a
preparatory step towards making `SaturationDetector` an official EPP
extension point.

- Moved files to `.../framework/plugins/utilizationdetector`
- Renamed `saturationdetector.go` to `detector.go`
- Renamed `saturationdetector_test.go` to `detector_test.go`
- Fixed imports

No functional changes are included in this commit.
