+++ pkg/epp/config/config.go
@@ -18,13 +18,13 @@ package config
 
 import (
 	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/datalayer"
-	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/saturationdetector"
+	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/saturationdetector/framework/plugins/utilizationdetector"
 	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/scheduling"
 )
 
 // Config is the configuration loaded from the text based configuration
 type Config struct {
 	SchedulerConfig          *scheduling.SchedulerConfig
-	SaturationDetectorConfig *saturationdetector.Config
+	SaturationDetectorConfig *utilizationdetector.Config
 	DataConfig               *datalayer.Config
 }
